tied in filter results to table and everything is working properly
