+++ bio-bid/src/components/CheckBoxFilters.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { GET_STUDIES } from '../queries';
-import { useQuery } from '@apollo/react-hooks';
+//import { GET_STUDIES } from '../queries';
+//import { useQuery } from '@apollo/react-hooks';
 
 /* STYLE IMPORT */
 import '../styles/filter.css';
@@ -9,131 +9,14 @@ import '../styles/filter.css';
 import { ReactComponent as CloseIcon } from '../images/close.svg';
 
 function CheckboxFilters(props) {
-    const [phase1, setPhase1] = useState(false);
-    const [phase2, setPhase2] = useState(false);
-    const [phase3, setPhase3] = useState(false);
-    const { data, loading, error } = useQuery(GET_STUDIES);
-    const Phases = {
-        none: 0,
-        phase1: 1,
-        phase2: 2,
-        phase3: 3
-    };
+    //const { data, loading, error } = useQuery(GET_STUDIES);
 
-    const [open, setOpen] = useState(false);
-    const [closed, setClosed] = useState(false);
-    const [active, setActive] = useState(false);
-    const Status = {
-        none: 0,
-        open: 1,
-        active: 2,
-        closed: 3
-    };
+    // const processFilter = (props) => {
 
-    const filterStudiesPhase = (phase) => {
-        if (data) {
-            switch (phase) {
-                case Phases.phase1:
-                    return data.studies.filter(study => study.phase === 1);
-
-                case Phases.phase2:
-                    return data.studies.filter(study => study.phase === 2);
-
-                case Phases.phase3:
-                    return data.studies.filter(study => study.phase === 3);
-
-                default:
-                    return [...data.studies];
-            }
-        }
-    }
-
-    const filterStudiesStatus = (status) => {
-        if (data) {
-            switch (status) {
-                case Status.open:
-                    return data.studies.filter(study => study.status === "Open");
-
-                case Status.active:
-                    return data.studies.filter(study => study.status === "Active");
-
-                case Status.closed:
-                    return data.studies.filter(study => study.status === "Closed");
-
-                default:
-                    return [...data.studies];
-            }
-        }
-    }
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        let phaseList = [];
-        let phaseTemp = [];
-        let statusList = [];
-        let statusTemp = [];
-
-        if (phase1) {
-            phaseTemp = filterStudiesPhase(Phases.phase1);
-            phaseList.push(...phaseTemp);
-
-        }
-
-        if (phase2) {
-            phaseTemp = filterStudiesPhase(Phases.phase2);
-            phaseList.push(...phaseTemp);
-        }
-
-        if (phase3) {
-            phaseTemp = filterStudiesPhase(Phases.phase3);
-            phaseList.push(...phaseTemp);
-        }
-
-        if (open) {
-            statusTemp = filterStudiesStatus(Status.open);
-            statusList.push(...statusTemp);
-
-        }
-
-        if (active) {
-            statusTemp = filterStudiesStatus(Status.active);
-            statusList.push(...statusTemp);
-        }
-
-        if (closed) {
-            statusTemp = filterStudiesStatus(Status.closed);
-            statusList.push(...statusTemp);
-        }
-
-        const finalList = [];
-
-        if (phaseList.length > 0) {
-            for (let i = 0; i < (phaseList.length + statusList.length); i++) {
-                if (phaseList[i] === statusList[i]) {
-                    finalList.push(phaseList[i])
-                }
-                else {
-                    const tempList = [phaseList[i], statusList[i]];
-                    finalList.push(...tempList);
-                }
-            }
-        }
-        else {
-            for (let k = 0; k < statusList.length; k++) {
-                finalList.push(statusList[k]);
-            }
-        }
-        const returnList = [];
-
-        for (let m = 0; m < finalList.length; m++) {
-            if (finalList[m] !== undefined && finalList[m] !== finalList[m + 1]) {
-                returnList.push(finalList[m]);
-            }
-        }
-        console.log(returnList)
-
-    }
-
+    //     setOpen(!open);
+    //     props.shadow();
+    
+      //}
     return (
         <div id='filter-wrapper' className={(props.open ? 'filter-open' : 'filter-closed')}>
             <div id='filter-header-container'>
@@ -147,20 +30,20 @@ function CheckboxFilters(props) {
                 <p className='cat-header'>Status:</p>
                 <input name='open'
                     type='checkbox'
-                    checked={open}
-                    onChange={() => setOpen(!open)} />
+                    checked={props.openStatus}
+                    onChange={() => props.setOpenStatus(!props.openStatus)} />
                 <label className='cat-item' >Open</label>
                 <br />
                 <input name='active'
                     type='checkbox'
-                    checked={active}
-                    onChange={() => setActive(!active)} />
+                    checked={props.active}
+                    onChange={() => props.setActive(!props.active)} />
                 <label className='cat-item' >Active</label>
                 <br />
                 <input name='closed'
                     type='checkbox'
-                    checked={closed}
-                    onChange={() => setClosed(!closed)} />
+                    checked={props.closed}
+                    onChange={() => props.setClosed(!props.closed)} />
                 <label className='cat-item' >Closed</label>
             </div>
             <hr className='filter-break' />
@@ -168,26 +51,26 @@ function CheckboxFilters(props) {
                 <p className='cat-header'>Phase:</p>
                 <input name='phase1'
                     type='checkbox'
-                    checked={phase1}
-                    onChange={() => setPhase1(!phase1)} />
+                    checked={props.phase1}
+                    onChange={() => props.setPhase1(!props.phase1)} />
                 <label className='cat-item' >I</label>
                 <br />
                 <input name='phase2'
                     type='checkbox'
-                    checked={phase2}
-                    onChange={() => setPhase2(!phase2)} />
+                    checked={props.phase2}
+                    onChange={() => props.setPhase2(!props.phase2)} />
                 <label className='cat-item' >II</label>
                 <br />
                 <input name='phase3'
                     type='checkbox'
-                    checked={phase3}
-                    onChange={() => setPhase3(!phase3)} />
+                    checked={props.phase3}
+                    onChange={() => props.setPhase3(!props.phase3)} />
                 <label className='cat-item' >III</label>
             </div>
             <hr className='filter-break' />
             <div style={{ display: 'flex', flexDirection: 'column', marginRight: '30px' }}>
-                <button className='final-filter-button' onClick={handleSubmit}>Apply Changes</button>
-                <button className='final-filter-button'>Clear all filters</button>
+                <button className='final-filter-button' onClick={props.handleFilterResults}>Apply Changes</button>
+                <button className='final-filter-button' onClick={props.clearFilters}>Clear all filters</button>
             </div>
         </div>
     );
