+++ bio-bid/src/components/bids.js
@@ -27,6 +27,30 @@ const Bids = (props) => {
   const [open, setOpen] = useState(false);
   const [studies, setStudies] = useState([]);
 
+  const [phase1, setPhase1] = useState(false);
+  const [phase2, setPhase2] = useState(false);
+  const [phase3, setPhase3] = useState(false);
+
+  const [openStatus, setOpenStatus] = useState(false);
+  const [closed, setClosed] = useState(false);
+  const [active, setActive] = useState(false);
+
+  const Phases = {
+    none: 0,
+    phase1: 1,
+    phase2: 2,
+    phase3: 3
+};
+
+const Status = {
+  none: 0,
+  open: 1,
+  active: 2,
+  closed: 3
+};
+
+
+
   const search = new JsSearch.Search('name');
 
   search.addIndex('name');
@@ -73,17 +97,129 @@ const Bids = (props) => {
 
   }
 
-  useEffect(() => {
+  const filterStudiesPhase = (phase) => {
+    if (data) {
+        switch (phase) {
+            case Phases.phase1:
+                return data.studies.filter(study => study.phase === 1);
 
+            case Phases.phase2:
+                return data.studies.filter(study => study.phase === 2);
+
+            case Phases.phase3:
+                return data.studies.filter(study => study.phase === 3);
+
+            default:
+                return [...data.studies];
+        }
+    }
+}
+
+const filterStudiesStatus = (status) => {
+    if (data) {
+        switch (status) {
+            case Status.open:
+                return data.studies.filter(study => study.status === "Open");
+
+            case Status.active:
+                return data.studies.filter(study => study.status === "Active");
+
+            case Status.closed:
+                return data.studies.filter(study => study.status === "Closed");
+
+            default:
+                return [...data.studies];
+        }
+    }
+}
+
+  const handleFilterResults = (event) => {
+    event.preventDefault();
+    let phaseList = [];
+    let phaseTemp = [];
+    let statusList = [];
+    let statusTemp = [];
+
+    if (phase1) {
+        phaseTemp = filterStudiesPhase(Phases.phase1);
+        phaseList.push(...phaseTemp);
+
+    }
+
+    if (phase2) {
+        phaseTemp = filterStudiesPhase(Phases.phase2);
+        phaseList.push(...phaseTemp);
+    }
+
+    if (phase3) {
+        phaseTemp = filterStudiesPhase(Phases.phase3);
+        phaseList.push(...phaseTemp);
+    }
+
+    if (openStatus) {
+        statusTemp = filterStudiesStatus(Status.open);
+        statusList.push(...statusTemp);
+
+    }
+
+    if (active) {
+        statusTemp = filterStudiesStatus(Status.active);
+        statusList.push(...statusTemp);
+    }
+
+    if (closed) {
+        statusTemp = filterStudiesStatus(Status.closed);
+        statusList.push(...statusTemp);
+    }
+
+    const finalList = [];
+
+    if (phaseList.length > 0) {
+        for (let i = 0; i < (phaseList.length + statusList.length); i++) {
+            if (phaseList[i] === statusList[i]) {
+                finalList.push(phaseList[i])
+            }
+            else {
+                const tempList = [phaseList[i], statusList[i]];
+                finalList.push(...tempList);
+            }
+        }
+    }
+    else {
+        for (let k = 0; k < statusList.length; k++) {
+            finalList.push(statusList[k]);
+        }
+    }
+    const returnList = [];
+
+    for (let m = 0; m < finalList.length; m++) {
+        if (finalList[m] !== undefined && finalList[m] !== finalList[m + 1]) {
+            returnList.push(finalList[m]);
+        }
+    }
+    setStudies(returnList);
+
+}
+
+const clearFilters = () => 
+{
+ setPhase1(false);
+ setPhase2(false);
+ setPhase3(false);
+ setOpenStatus(false);
+ setClosed(false);
+ setActive(false);
+  setStudies(data.studies);
+}
+
+  useEffect(() => {
       if (data !== undefined) {
 
           search.addDocuments(data.studies);
 
           if (studies.length <= 0 && value.length <= 0)
               setStudies(data.studies);
-
       }
-
   });
 
   return (
@@ -141,7 +277,24 @@ const Bids = (props) => {
           </tbody>
         </Table>
       </div>
-      <CheckBoxFilters process={processFilter} open={open} />
+      <CheckBoxFilters 
+        process={processFilter} 
+        open={open} 
+        handleFilterResults={handleFilterResults}
+        clearFilters={clearFilters}
+        phase1={phase1}
+        setPhase1={setPhase1}
+        phase2={phase2}
+        setPhase2={setPhase2}
+        phase3={phase3}
+        setPhase3={setPhase3}
+        openStatus={openStatus}
+        setOpenStatus={setOpenStatus}
+        active={active}
+        setActive={setActive}
+        closed={closed}
+        setClosed={setClosed}
+      />
     </div>
   );
 }
